pyproject.toml
[tool.poetry]
name = "wordCountBot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
praw = "^7.5.0"
rich = "^11.0.0"
python-box = "^5.4.1"
python-dotenv = "^0.19.2"
dateparser = "^1.1.0"
humanize = "^3.13.1"
Pygments = "^2.11.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
